refactor : PieceType 패키지 이동
